add utils module
Move all generic utility functions there.
