+++ xpublish/utils.py
@@ -1,6 +1,8 @@
-"""Utility functions for printing version information.
+"""Utility functions for internal use only.
 
-Adapted from xarray/util/print_versions.py
+The utility functions for printing version information have been adapted from
+xarray/util/print_versions.py
+
 """
 
 import locale
@@ -9,6 +11,7 @@ import platform
 import struct
 import subprocess
 import sys
+import time
 
 
 def get_sys_info():
@@ -78,3 +81,15 @@ def netcdf_and_hdf5_versions():
         except ImportError:
             pass
     return [('libhdf5', libhdf5_version), ('libnetcdf', libnetcdf_version)]
+
+
+class CostTimer:
+    """ Context manager to measure wall time """
+
+    def __enter__(self):
+        self._start = time.perf_counter()
+        return self
+
+    def __exit__(self, *args):
+        end = time.perf_counter()
+        self.time = end - self._start
